Added a command to admin that lets you add a pub to a certain area. Fixed a bug allowing any username, password entered in for the admin to be correct.

+++ src/main/java/ie/atu/interface1/pubOverview.java
@@ -22,7 +22,7 @@ public class pubOverview extends pubSelect {
     public static double calcAverage(double[] rating) {
         int totalSum = 0;
         for (double num : rating) {
-            totalSum += num;
+            totalSum += (int) num;
         }
         return (double) totalSum / rating.length;
     }
